% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitall.r
\name{fitall}
\alias{fitall}
\title{Fit basic statistical models and summarize to Maths data}
\usage{
fitall(obj, fit_type = c("lm", "summarise_model", "summarise_data"))
}
\arguments{
\item{obj}{An object of class \code{Analysis} from \code{\link{MathsData}}}

\item{fit_type}{The linear regression model(\code{lm}), summarize the model and data (\code{summarise_model}),(\code{summarise_data})}
}
\value{
Returns a list of class \code{db_fit} which includes the model details as well as the data set and fit type used or summary of data or summary of model
}
\description{
Fit basic statistical models and summarize to Maths data
}
\examples{
\dontrun{
path = system.file("extdata","Maths_ED.xlsx", package = "Analysis")
data = Analysis::MathsData(path)
ans1 = Analysis::fitall(data, 'lm')
}
}
