% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathData.R
\name{MathsData}
\alias{MathsData}
\title{Loading Maths data}
\usage{
MathsData(path)
}
\arguments{
\item{path}{Give path in \code{MathsData} function}
}
\value{
A \code{\link[tibble]{tibble}} which contains 3 columns Gender,Maths scores and VLE use
}
\description{
Loading Maths data
}
\examples{

path = system.file("extdata","Maths_ED.xlsx", package = "Analysis")
data = Analysis::MathsData(path)

}
