% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotanime.db_fit.R
\name{plotanime.db_fit}
\alias{plotanime.db_fit}
\title{Animate Plot packageMaths output}
\usage{
plotanime.db_fit(y)
}
\arguments{
\item{y}{Output from the \code{fitall} function}
}
\value{
Nothing: just a nice anime plot
}
\description{
Animate Plot packageMaths output
}
\examples{
\dontrun{
path = system.file("extdata","Maths_ED.xlsx", package = "packageMaths")
data = Analysis::MathsData(path)
ans1 = Analysis::fitall(data, 'lm')
Analysis::plotanime.db_fit(ans1)
}
}
