% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotR.R
\name{plotR}
\alias{plotR}
\title{Plot packageMaths output}
\usage{
plotR(x, VLE_Use = pretty(x$data$VLE_Use, n = 100), ...)
}
\arguments{
\item{x}{Output from the \code{fit} function}

\item{VLE_Use}{An optional time grid for VLE_use over which to produce fitted values of the model}

\item{...}{Other arguments to plot (not currently implemented)}


}
\value{
Nothing: just a nice plot
}
\description{
Plot packageMaths output
}
\examples{
\dontrun{
path = system.file("extdata","Maths_ED.xlsx", package = "Analysis")
data = Analysis::MathsData(path)
ans1 = Analysis::fitall(data, 'lm')
Analysis::plotR(ans1)
}
}
