% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Example.R
\name{run_Example}
\alias{run_Example}
\title{Run Shiny application for the Analysis package}
\usage{
run_Example()
}
\value{
Runs the Shiny application for the Analysis package
}
\description{
Run Shiny application for the Analysis package
}
